Do not use nested Buffer suite allocations
PhotoFiltre crashes with an access violation when it tries to free one
of the buffers.
Using new for the PNG row pointers also reduces the amount to code
required to allocate the memory.

An unique_ptr is not used because LibPNG uses setjmp for error
handling.
The MSDN documentation for setjmp warns that it is not guaranteed to
work with the automatic C++ object destruction.