Always show a folder prompt for multiple output images
When running as a repeat effect the images in the last used output
folder would be silently overwritten, which is a data loss issue.
Always showing a folder browser prompt to the user makes it clear that
the filter has produced multiple output images, and will not be
changing the active layer.