Convert high bit depth images to 8-bits-per-channel
There is some kind of issue with the WIC conversion that causes 16-bit
images to be displayed as corrupted.
The pixel values look correct in the debugger, but the final QImage is
corrupt.

This change only affects images that are loaded from a file or a PNG
that is loaded from a memory buffer, DIB images are always
8-bits-per-channel.