Use planar order when copying data to the host
Some hosts (e.g. PaintShop Pro 2021) do not include the layer
transparency in the FilterRecord outColumnBytes value when a plugin
uses interleaved channel processing.
This causes the pixel data to be corrupted if the layer supports
editing transparency.

The fix is to process each plane individually, it is slower but should
work correctly for all host applications.