Support hosts that ignore the plug-in error codes
A number of popular host applications silently ignore the standard
error codes that Photoshop-compatible plug-ins can return to the host
instead of showing the user an appropriate error message.
Examples include: Affinity Photo version 1.9, IrfanView and XnView.

To fix this the plug-in will now show generic error messages to the
user for the standard error codes.
Additionally the plug-in now uses exceptions more widely for reporting
error conditions as this can provide the user with more descriptive
error messages, this also greatly simplifies the error handling code.

Fixes #7