Use the GMIC 8bf image format when importing
This allows G'MIC-Qt to write a planar image with tiling for the images
that are copied into the active layer.
The planar with tiling image format reduces memory consumption when
loading most images.

The images that are saved to an external file will be written in an
interleaved pixel format, and converted to PNG when saving the output
image.