Improve the integration with the Actions palette
The plugin will now show the G'MIC-Qt effect name and input layer mode
when the user expands the "G'MIC-Qt" entry in the actions palette.
The G'MIC-Qt settings are now saved by the plugin, this allows multiple
G'MIC-Qt effects to be used in a single automation script.

Note that G'MIC-Qt will show its progress dialog even when the user has
requested the plugin to not show a dialog, this is done so that the
user sees the current progress for effects that take a while to render.