Add the G'MIC command name to the PNG file names
The G'MIC-Qt 8bf host currently writes an image that uses a file name
with a timestamp (and image sequence number, if required).
The 8bf plug-in will add the G'MIC command name in front of the
original file name when writing the external PNG files.

The final image file name will use the following format:
<G'MIC command name>_<original file name>

According to the G'MIC Language Reference command names are restricted
to a subset of 7-bit US-ASCII: letters, numbers and underscores.
These restrictions make the command name safe to use in an OS file
name.