Add an ARM64 platform target
The ARM64 version of Windows 10 can run the X86 version of G'MIC-Qt in
an emulated environment.
The ARM64 version of Windows 11 can run the X64 version of G'MIC-Qt in
an emulated environment.

Unfortunately, a native ARM64 build of G'MIC-Qt is not currently
possible.
MinGW-w64 does not have ARM64 versions of the required dependencies,
and some of those dependencies fail to build when cross-compiling with
Visual Studio 2022 and vcpkg.