Get the image depth from the FilterRecord depth field
It would be valid for a host to implement 16-bit support by using
plugInModeRGBColor with the depth field set to 16.

The image mode will still be used as a fallback for detecting the image
depth for hosts that do not set the depth field.
For example, Photoshop versions prior to 5.0.