Support color management for the G'MIC-Qt preview
This improves the appearance of the preview image for all image modes.
Color management will only be used if the 8bf plug-in host provides the
image color profile to its 8bf plug-ins.

The 8bf plug-in will also provide G'MIC-Qt with the color profile for
the primary display, this will be used as a fallback if G'MIC-Qt
cannot retrieve the color profile for the display it is shown on.