Fix the error handling when preallocating a file
Because preallocating a file is a performance optimization, we only
throw an exception for errors that indicate there is not enough space.
If SetFileInformationByHandle fails for any other reason, we write the
file without preallocating it.

This also fixes a compatibility issue between the plugin and Wine.
As of Wine 10.0 RC3 their SetFileInformationByHandle implementation
does not support the FileAllocationInfo class type.